requirements.txt
requests==2.31.0
python-dotenv==1.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
flask==3.0.0
bootstrap-flask==2.3.3
